开发点赞投币收藏功能（未完成）

+++ cilicili-back/src/main/java/com/cilicili/entity/User.java
@@ -3,6 +3,8 @@ package com.cilicili.entity;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @TableName("user")
 public class User {
@@ -10,12 +12,11 @@ public class User {
     private String username;
     private String password;
     private String nickname;
-//    头像URL
-    private String avatar;
-    private String signature;
+    private String avatar;      // 头像URL
+    private String signature;   // 签名
     private String email;
     private String phone;
-    private Integer coin;
+    private BigDecimal coin;
     private Byte status;
     private Byte role;
     private String created_at;
